More refactoring & Layouts
