--- src/screens/ContactCreateScreen/ContactCreateScreen.tsx
@@ -1,10 +1,0 @@
-import { View, Text } from 'react-native';
-import React from 'react';
-
-const ContactCreateScreen = () => (
-	<View>
-		<Text>ContactCreateScreen</Text>
-	</View>
-);
-
-export default ContactCreateScreen;
