+++ src/screens/SaveScreen/Background/Background.styles.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		position: 'absolute',
+		height: '100%',
+		width: '115%'
+	}
+});
