+++ src/screens/WalletAssetsScreen/WalletAssetsScreen.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { View } from 'react-native';
 import { useTheme, useTokens } from '@hooks';
 import { Modal, ModalReusables } from '@components';
+import { BasicLayout } from '@layouts';
 import AssetList from './AssetList/AssetList';
 import ValueBox from './ValueBox/ValueBox';
 import AssetListEmpty from './AssetListEmpty/AssetListEmpty';
@@ -14,7 +15,7 @@ const WalletAssetsScreen = () => {
 
 	return (
 		<>
-			<View style={[styles.container, { backgroundColor: colors.detail4 }]}>
+			<BasicLayout hideSafeAreaView bg="detail4">
 				<ValueBox {...{ balance }} />
 				<View style={[styles.assetListContainer, { backgroundColor: colors.background1 }]}>
 					{tokens && tokens.length > 0 ? (
@@ -23,7 +24,7 @@ const WalletAssetsScreen = () => {
 						<AssetListEmpty onPress={() => setAddFundsVisible(true)} />
 					)}
 				</View>
-			</View>
+			</BasicLayout>
 			<Modal isVisible={addFundsVisible} onDismiss={() => setAddFundsVisible(false)}>
 				<ModalReusables.ComingSoon onDismiss={() => setAddFundsVisible(false)} />
 			</Modal>
